refactor: 어드민 리팩토링
