+++ src/main/java/roomescape/controller/admin/AdminController.java
@@ -4,7 +4,6 @@ import jakarta.validation.Valid;
 import java.net.URI;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -29,21 +28,6 @@ public class AdminController {
     public AdminController(ReservationService reservationService, ReservationMemberService reservationMemberService) {
         this.reservationService = reservationService;
         this.reservationMemberService = reservationMemberService;
-    }
-
-    @GetMapping("/reservation")
-    public String reservationPage() {
-        return "admin/reservation-new";
-    }
-
-    @GetMapping("/time")
-    public String reservationTimePage() {
-        return "admin/time";
-    }
-
-    @GetMapping("/theme")
-    public String themePage() {
-        return "admin/theme";
     }
 
     @PostMapping("/reservations")
